+++ examples/src/main/java/com/mapbox/navigation/examples/core/replay/HistoryFilesActivity.kt
@@ -11,7 +11,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.appbar.CollapsingToolbarLayout
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 import com.google.android.material.snackbar.Snackbar
-import com.mapbox.navigation.core.history.MapboxHistoryReader
+import com.mapbox.navigation.core.replay.history.ReplayHistoryEventStream
 import com.mapbox.navigation.examples.core.R
 
 @SuppressLint("HardwareIds")
@@ -20,7 +20,7 @@ class HistoryFilesActivity : AppCompatActivity() {
     companion object {
         const val REQUEST_CODE: Int = 123
         const val EXTRA_HISTORY_FILE_DIRECTORY = "EXTRA_HISTORY_FILE_DIRECTORY"
-        var selectedHistory: MapboxHistoryReader? = null
+        var selectedHistory: ReplayHistoryEventStream? = null
             private set
     }
 
@@ -53,8 +53,8 @@ class HistoryFilesActivity : AppCompatActivity() {
 
         val historyFileDirectory = intent.extras?.getString(EXTRA_HISTORY_FILE_DIRECTORY)
         filesViewController = HistoryFilesViewController(historyFileDirectory)
-        filesViewController.attach(this, viewAdapter) { historyDataResponse ->
-            if (historyDataResponse == null) {
+        filesViewController.attach(this, viewAdapter) { replayHistoryEventStream ->
+            if (replayHistoryEventStream == null) {
                 Snackbar.make(
                     recyclerView,
                     getString(R.string.history_failed_to_load_item),
@@ -64,7 +64,7 @@ class HistoryFilesActivity : AppCompatActivity() {
                     null
                 ).show()
             } else {
-                selectedHistory = historyDataResponse
+                selectedHistory = replayHistoryEventStream
                 finish()
             }
         }