+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/history/MapboxHistoryReader.kt
@@ -1,50 +1,55 @@
 package com.mapbox.navigation.core.history
 
+import androidx.annotation.VisibleForTesting
 import com.mapbox.navigation.core.history.model.HistoryEvent
 import com.mapbox.navigation.core.history.model.HistoryEventMapper
 import com.mapbox.navigator.HistoryReader
+import com.mapbox.navigator.HistoryReaderInterface
 
 /**
  * Allows you to read history files previously saved by [MapboxHistoryRecorder].
  * All files in the [MapboxHistoryRecorder.fileDirectory] can be read with this reader.
- *
- * @param filePath absolute path to a file containing the native history file.
  */
-class MapboxHistoryReader(
-    val filePath: String
+class MapboxHistoryReader @VisibleForTesting internal constructor(
+    private val nativeHistoryReader: HistoryReaderInterface,
+    private val historyEventMapper: HistoryEventMapper
 ) : Iterator<HistoryEvent> {
 
-    private val nativeHistoryReader = HistoryReader(filePath)
-    private val historyEventMapper = HistoryEventMapper()
+    /**
+     * @param filePath absolute path to a file containing the native history file.
+     */
+    constructor(filePath: String) : this(
+        HistoryReader(filePath),
+        HistoryEventMapper()
+    )
 
-    private var hasNext: Boolean = false
     private var next: HistoryEvent? = null
 
     init {
-        hasNext = loadNext()
+        loadNext()
     }
 
     /**
      * Returns `true` if the iteration has more elements.
      */
-    override fun hasNext(): Boolean = hasNext
+    override fun hasNext(): Boolean = next != null
 
     /**
      * Returns the next element in the iteration.
      *
      * @throws NullPointerException when [hasNext] is false
      */
+    @Throws(NullPointerException::class)
     override fun next(): HistoryEvent = next!!.also {
-        hasNext = loadNext()
+        loadNext()
     }
 
-    private fun loadNext(): Boolean {
+    private fun loadNext() {
         val historyRecord = nativeHistoryReader.next()
         next = if (historyRecord != null) {
             historyEventMapper.map(historyRecord)
         } else {
             null
         }
-        return next != null
     }
 }