+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/replay/MapboxReplayer.kt
@@ -10,9 +10,9 @@ import com.mapbox.android.core.location.LocationEngineCallback
 import com.mapbox.android.core.location.LocationEngineProvider
 import com.mapbox.android.core.location.LocationEngineResult
 import com.mapbox.navigation.core.replay.history.ReplayEventBase
+import com.mapbox.navigation.core.replay.history.ReplayEventBuffer
 import com.mapbox.navigation.core.replay.history.ReplayEventSimulator
 import com.mapbox.navigation.core.replay.history.ReplayEventUpdateLocation
-import com.mapbox.navigation.core.replay.history.ReplayEvents
 import com.mapbox.navigation.core.replay.history.ReplayEventsObserver
 import com.mapbox.navigation.core.replay.route.ReplayRouteMapper
 import com.mapbox.navigation.core.replay.route.ReplayRouteOptions
@@ -24,8 +24,8 @@ import java.util.Collections.singletonList
 @UiThread
 class MapboxReplayer {
 
-    private val replayEvents = ReplayEvents(mutableListOf())
-    private val replayEventSimulator = ReplayEventSimulator(replayEvents)
+    private val replayEventBuffer = ReplayEventBuffer()
+    private val replayEventSimulator = ReplayEventSimulator(replayEventBuffer)
 
     private val replayEventsObservers: MutableSet<ReplayEventsObserver> = mutableSetOf()
 
@@ -36,9 +36,18 @@ class MapboxReplayer {
      * @param events the events to be replayed.
      * @return [MapboxReplayer]
      */
-    fun pushEvents(events: List<ReplayEventBase>): MapboxReplayer {
-        this.replayEvents.events.addAll(events)
-        return this
+    fun pushEvents(events: List<ReplayEventBase>) = apply {
+        replayEventBuffer.pushEvents(events)
+    }
+
+    /**
+     * Attach a stream of events to be replayed. There can only be one stream at a time. When a new
+     * stream is attached, the previous stream will be closed automatically.
+     *
+     * This method for pushing events, is more memory efficient than pushing large event lists.
+     */
+    fun attachStream(eventStream: ReplayEventStream) = apply {
+        replayEventBuffer.attachStream(eventStream)
     }
 
     /**
@@ -48,7 +57,7 @@ class MapboxReplayer {
     fun clearEvents() {
         stop()
         seekTo(0.0)
-        replayEvents.events.clear()
+        replayEventBuffer.clear()
     }
 
     /**
@@ -84,7 +93,9 @@ class MapboxReplayer {
      * registered via [registerObserver]
      */
     fun play() {
+        replayEventBuffer.bufferEvents()
         replayEventSimulator.launchSimulator { replayEvents ->
+            replayEventBuffer.bufferEvents()
             replayEventsObservers.forEach { it.replayEvents(replayEvents) }
         }
     }
@@ -121,10 +132,10 @@ class MapboxReplayer {
      * Use this function to play the first location received from your [LocationEngine].
      */
     fun playFirstLocation() {
-        val firstUpdateLocationIndex = replayEvents.events.indexOfFirst { replayEvent ->
+        val firstUpdateLocationIndex = replayEventBuffer.events.indexOfFirst { replayEvent ->
             replayEvent is ReplayEventUpdateLocation
         }
-        replayEvents.events.getOrNull(firstUpdateLocationIndex)?.let { replayEvent ->
+        replayEventBuffer.events.getOrNull(firstUpdateLocationIndex)?.let { replayEvent ->
             seekTo(replayEvent)
             val replayEvents = singletonList(replayEvent)
             replayEventsObservers.forEach { it.replayEvents(replayEvents) }
@@ -160,9 +171,9 @@ class MapboxReplayer {
      * @return the duration in seconds
      */
     fun durationSeconds(): Double {
-        val firstEvent = replayEvents.events.firstOrNull()
+        val firstEvent = replayEventBuffer.events.firstOrNull()
             ?: return 0.0
-        val lastEvent = replayEvents.events.last()
+        val lastEvent = replayEventBuffer.events.last()
         return lastEvent.eventTimestamp - firstEvent.eventTimestamp
     }
 
@@ -170,7 +181,7 @@ class MapboxReplayer {
      * The time of an event, relative to the duration of the replay.
      */
     fun eventSeconds(eventTimestamp: Double): Double {
-        val firstEvent = replayEvents.events.firstOrNull()
+        val firstEvent = replayEventBuffer.events.firstOrNull()
             ?: return 0.0
         return eventTimestamp - firstEvent.eventTimestamp
     }
@@ -194,10 +205,10 @@ class MapboxReplayer {
      * @param replayTime time in seconds between 0.0 to [durationSeconds]
      */
     fun seekTo(replayTime: Double) {
-        val firstEventTime = replayEvents.events.firstOrNull()?.eventTimestamp
+        val firstEventTime = replayEventBuffer.events.firstOrNull()?.eventTimestamp
             ?: return
         val offsetTime = replayTime + firstEventTime
-        val indexOfEvent = replayEvents.events
+        val indexOfEvent = replayEventBuffer.events
             .indexOfFirst { offsetTime <= it.eventTimestamp }
         check(indexOfEvent >= 0) {
             "Make sure your replayTime is less than replayDurationSeconds " +
@@ -214,7 +225,7 @@ class MapboxReplayer {
      * @throws IllegalStateException if [replayEvent] was not pushed
      */
     fun seekTo(replayEvent: ReplayEventBase) {
-        val indexOfEvent = replayEvents.events.indexOf(replayEvent)
+        val indexOfEvent = replayEventBuffer.events.indexOf(replayEvent)
         check(indexOfEvent >= 0) { "You must first pushEvents and then seekTo an event" }
 
         replayEventSimulator.seekTo(indexOfEvent)